domain/src/main/java/com/example/domain/model/Feed.java
package com.example.domain.model;

public class Feed {
    String title;
    String description;
    String detailUrl;
    String thumbnailUrl;

    public Feed(String title, String description, String detailUrl, String thumbnailUrl) {
        this.title = title;
        this.description = description;
        this.detailUrl = detailUrl;
        this.thumbnailUrl = thumbnailUrl;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getDetailUrl() {
        return detailUrl;
    }

    public void setDetailUrl(String detailUrl) {
        this.detailUrl = detailUrl;
    }

    public String getThumbnailUrl() {
        return thumbnailUrl;
    }

    public void setThumbnailUrl(String thumbnailUrl) {
        this.thumbnailUrl = thumbnailUrl;
    }
}
